backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiohttp>=3.12.12",
    "deepgram-sdk>=4.3.0",
    "fastapi[standard]>=0.115.12",
    "python-dotenv>=1.1.0",
    "uvicorn>=0.34.3",
]
